Made home screen a calculator
This is a more natural format for the home screen in my opinion since 95%
of the time the app will be used it will be done so in order to add
their expenses.  I also updated the money values to use BigDecimal to
appropriately round everything to a hundredths digit precision.